requirements.txt
uvicorn~=0.29.0
fastapi~=0.115.8
requests~=2.32.5
pillow~=10.3.0
pydantic~=2.7.1
typing_extensions~=4.11.0
torch~=2.5.0+cu118
transformers~=4.57.1
trl~=0.24.0
numpy~=2.2.0
datasets~=4.3.0
peft~=0.17.1
fsspec~=2024.3.1
tqdm~=4.67.1
accelerate~=1.11.0
safetensors~=0.6.2
psutil~=6.1.1
PyYAML~=6.0.1
gradio~=5.9.1
matplotlib~=3.8.4
packaging~=24.0
wandb~=0.22.2
scikit-image~=0.25.2